Update imports and documentation for settings views since they were recently moved
